Add history package into telepathy package and refactor code
